첫 페이지 getserversiderendering 중복 요청되는거 수정
